Create Event abstract class
