+++ backend/config/packages/doctrine.yaml
@@ -27,3 +27,9 @@ doctrine:
                 dir: '%kernel.project_dir%/packages/Fishbowl/src/Entity'
                 prefix: 'App\Fishbowl\Entity'
                 alias: App\Fishbowl
+            App\WorldCafe:
+                is_bundle: false
+                type: annotation
+                dir: '%kernel.project_dir%/packages/WorldCafe/src/Entity'
+                prefix: 'App\Fishbowl\Entity'
+                alias: App\Fishbowl
