+++ backend/packages/WorldCafe/config/services.yaml
@@ -1,0 +1,17 @@
+# This file is part of the Stooa codebase.
+#
+# (c) 2020 - present Runroom SL
+#
+# For the full copyright and license information, please view the LICENSE
+# file that was distributed with this source code.
+
+services:
+    _defaults:
+        autowire: true
+        autoconfigure: true
+
+    App\WorldCafe\:
+        resource: ../src/
+        exclude:
+            - ../src/DependencyInjection/
+            - ../src/Entity/
