+++ backend/packages/WorldCafe/src/Entity/WorldCafe.php
@@ -1,0 +1,64 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Stooa codebase.
+ *
+ * (c) 2020 - present Runroom SL
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\WorldCafe\Entity;
+
+use ApiPlatform\Core\Annotation\ApiResource;
+use App\Core\Model\Event;
+use App\WorldCafe\Repository\WorldCafeRepository;
+use Doctrine\ORM\Mapping as ORM;
+use Ramsey\Uuid\Doctrine\UuidGenerator;
+use Ramsey\Uuid\Uuid;
+use Ramsey\Uuid\UuidInterface;
+use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\Serializer\Annotation\Groups;
+
+/**
+ * @ApiResource(
+ *     normalizationContext={"groups"={"event:write"}},
+ *     denormalizationContext={"groups"={"event:write"}},
+ * )
+ * @UniqueEntity(fields={"slug"})
+ * @ORM\Entity(repositoryClass=WorldCafeRepository::class)
+ */
+final class WorldCafe extends Event
+{
+    /**
+     * @Groups({"event:read"})
+     * @ORM\Id
+     * @ORM\Column(type="uuid", unique=true)
+     * @ORM\GeneratedValue(strategy="CUSTOM")
+     * @ORM\CustomIdGenerator(class=UuidGenerator::class)
+     */
+    private ?UuidInterface $id = null;
+
+    public function __toString(): string
+    {
+        $uid = $this->getId();
+        $stringUid = null !== $uid ? ' (' . $uid->toString() . ')' : '';
+
+        return ($this->getName() ?? '') . $stringUid;
+    }
+
+    public function getId(): ?UuidInterface
+    {
+        return $this->id;
+    }
+
+    public function setId(string $id): self
+    {
+        $this->id = Uuid::fromString($id);
+
+        return $this;
+    }
+}
